[nextest-runner] start refactoring the displayer
The displayer's getting too big -- start splitting out chunks of it. Do
progress reporting first.
