+++ nextest-runner/src/reporter/displayer/mod.rs
@@ -1,0 +1,10 @@
+// Copyright (c) The nextest Contributors
+// SPDX-License-Identifier: MIT OR Apache-2.0
+
+//! The displayer for human-friendly output.
+
+mod duration;
+mod imp;
+mod progress;
+
+pub use imp::*;
