+++ nextest-runner/src/reporter/helpers.rs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
 use bstr::ByteSlice;
+use owo_colors::Style;
 
 /// Given a slice, find the index of the point at which highlighting should end.
 ///
@@ -19,6 +20,31 @@ pub fn highlight_end(slice: &[u8]) -> usize {
         }
         // No newline found, so highlight the entire slice.
         None => slice.len(),
+    }
+}
+
+#[derive(Debug, Default)]
+pub(super) struct Styles {
+    pub(super) is_colorized: bool,
+    pub(super) count: Style,
+    pub(super) pass: Style,
+    pub(super) retry: Style,
+    pub(super) fail: Style,
+    pub(super) skip: Style,
+    pub(super) script_id: Style,
+    pub(super) list_styles: crate::list::Styles,
+}
+
+impl Styles {
+    pub(super) fn colorize(&mut self) {
+        self.is_colorized = true;
+        self.count = Style::new().bold();
+        self.pass = Style::new().green().bold();
+        self.retry = Style::new().magenta().bold();
+        self.fail = Style::new().red().bold();
+        self.skip = Style::new().yellow().bold();
+        self.script_id = Style::new().blue().bold();
+        self.list_styles.colorize();
     }
 }
